Update feature(ceo): profile setting - Fix feature(ceo): fix error of edit user information that image can be other type and username can be duplicated.

+++ src/java/controller/ceo/CeoProfileServlet.java
@@ -5,8 +5,6 @@
 package controller.ceo;
 
 import dal.CeoDAO;
-import model.Role;
-import model.Staff;
 import java.io.IOException;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.MultipartConfig;
@@ -16,12 +14,10 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import jakarta.servlet.http.Part;
 import java.io.File;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
+import model.Staff;
+import org.apache.tomcat.dbcp.pool2.PoolUtils;
 import util.AccountValidation;
 
 /**
@@ -29,7 +25,7 @@ import util.AccountValidation;
  * @author Long
  */
 @MultipartConfig
-public class EditStaffInfo extends HttpServlet {
+public class CeoProfileServlet extends HttpServlet {
 
     private boolean deleteFile(String relativeFilePath) {
         try {
@@ -96,19 +92,6 @@ public class EditStaffInfo extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-        String acccountId = request.getParameter("uid");
-        CeoDAO aDao = new CeoDAO();
-        try {
-            int uid = Integer.parseInt(acccountId);
-            Staff a = aDao.getStaffById(uid);
-            request.setAttribute("staff", a);
-            List<Role> roles = aDao.getAllRoles();
-            request.setAttribute("roles", roles);
-            request.getRequestDispatcher("ceo/editStaffInfo.jsp").forward(request, response);
-        } catch (NumberFormatException e) {
-
-        }
-
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
@@ -140,40 +123,52 @@ public class EditStaffInfo extends HttpServlet {
         HttpSession session = request.getSession();
         CeoDAO aDao = new CeoDAO();
         AccountValidation validate = new AccountValidation();
-        int id = Integer.parseInt(request.getParameter("id"));
+        Staff currentAccount = (Staff) session.getAttribute("staff");
+
+//        List<String> errorPassMess = new ArrayList<>();
+//        String action = request.getParameter("action");
+//        if (action != null) {
+//            try {
+//                String currentPassword = request.getParameter("oldPassword");
+//                String newPassword = request.getParameter("newPassword");
+//
+//                if (!validate.hashPassword(currentPassword).equals(currentAccount.getPassword())) {
+//                    errorPassMess.add("Your current password is incorrect");
+//                } else if (!validate.checkHashOfPassword(newPassword)) {
+//                    errorPassMess.add("Password must be at least 8 characters long and include an uppercase letter, "
+//                            + "a lowercase letter, a number, a special character and not contains space characters.");
+//                } else if (validate.hashPassword(newPassword).equals(currentAccount.getPassword())) {
+//                    errorPassMess.add("Your new password matches your old password");
+//                }
+//                if (!errorPassMess.isEmpty()) {
+//                    request.setAttribute("errorPassMess", errorPassMess);
+//                    request.getRequestDispatcher("ceo/ceoProfile.jsp").forward(request, response);
+//                    return;
+//                }
+//                aDao.updatePassword(currentAccount.getId(), validate.hashPassword(newPassword));
+//                currentAccount = aDao.getStaffById(currentAccount.getId());
+//                request.setAttribute("successMess", "Change password successfully. ");
+//                session.setAttribute("staff", currentAccount);
+//            } catch (Exception e) {
+//                System.out.println(e);
+//            }
+//            request.getRequestDispatcher("ceo/ceoProfile.jsp").forward(request, response);
+//            return;
+//        }
         String username = validate.normalizeInput(request.getParameter("username"));
-        String firstName = validate.normalizeInput(request.getParameter("firstName"));
-        String lastName = validate.normalizeInput(request.getParameter("lastName"));
-        String phone = request.getParameter("phone").trim();
-        String email = request.getParameter("email").trim();
+        String firstName = validate.normalizeInput(request.getParameter("firstname"));
+        String lastName = validate.normalizeInput(request.getParameter("lastname"));
         String address = validate.normalizeInput(request.getParameter("address"));
-        String role = request.getParameter("role");
         String gender = request.getParameter("gender");
-        String dob = request.getParameter("dob");
-        String salary = request.getParameter("salary");
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        LocalDate date = LocalDate.parse(dob, formatter);
-        BigDecimal bigDecimalValue = new BigDecimal(salary);
         Part imagePart = request.getPart("image");
         String fileType = imagePart.getContentType();
         List<String> errorMessages = new ArrayList<>();
-        List<Role> roles = aDao.getAllRoles();
-        request.setAttribute("roles", roles);
-
         // Xác thực dữ liệu
         if (validate.normalizeInput(username).isEmpty()) {
             errorMessages.add("Username is required.");
         }
-        if (aDao.getStaffByUsername(username) != null && id != aDao.getStaffByUsername(username).getId()) {
+        if (aDao.getStaffByUsername(username) != null && currentAccount.getId() != aDao.getStaffByUsername(username).getId()) {
             errorMessages.add("Duplicate username.");
-        }
-
-        if (!validate.isValidEmail(email)) {
-            errorMessages.add("Invalid email format.");
-        }
-
-        if (aDao.getStaffByEmail(email) != null && id != aDao.getStaffByEmail(email).getId()) {
-            errorMessages.add("Duplicate email.");
         }
 
         if (!validate.isAlphabetic(validate.normalizeInput(firstName))) {
@@ -183,71 +178,38 @@ public class EditStaffInfo extends HttpServlet {
         if (!validate.isAlphabetic(validate.normalizeInput(lastName))) {
             errorMessages.add("Invalid last name");
         }
-
-        if (dob == null || dob.isEmpty()) {
-            errorMessages.add("Date of Birth is required.");
-        } else {
-            try {
-                // 2. Kiểm tra định dạng và ngày hợp lệ
-                LocalDate dateOfBirth = LocalDate.parse(dob, formatter);
-                LocalDate today = LocalDate.now();
-
-                // 3. Ngày không được ở tương lai
-                if (dateOfBirth.isAfter(today)) {
-                    errorMessages.add("Date of Birth cannot be in the future.");
-                }
-            } catch (Exception e) {
-                errorMessages.add("Invalid date.");
-            }
-        }
-
-        if (!validate.isValidPhone(phone)) {
-            errorMessages.add("Invalid phone number.");
-        }
-        if (aDao.getStaffByPhone(phone) != null && id != aDao.getStaffByPhone(phone).getId()) {
-            errorMessages.add("Duplicate phone number.");
-        }
-
-        double w = 0;
-        try {
-            w = Double.parseDouble(salary);
-            if (w < 0) {
-                errorMessages.add("Salary cannot be negative.");
-            }
-        } catch (NumberFormatException e) {
-            errorMessages.add("Salary must be a valid number.");
-        }
         String image = (imagePart != null && imagePart.getSize() > 0 ? getAndSaveImg(imagePart) : null);
 
         if (imagePart.getSize() > 1024 * 1024 * 5) {
             errorMessages.add("Image must be < 5mb");
         }
-        if(image!=null && !validate.isValidateImage(image)) {
+        if (image != null && !validate.isValidateImage(image)) {
             errorMessages.add("Image must be .jpg, .jpeg, .png");
         }
-
         // Nếu có lỗi, trả về trang trước đó với thông báo lỗi
+
         if (!errorMessages.isEmpty()) {
             request.setAttribute("errorMessages", errorMessages);
-            request.setAttribute("staff", new Staff(id, username, "", image, email, firstName, lastName, gender, date, phone, address, bigDecimalValue,
-                    id, LocalDateTime.MIN, new Role(aDao.getRoleIdByName(role), role)));
-            request.getRequestDispatcher("ceo/editStaffInfo.jsp").forward(request, response);
+            request.getRequestDispatcher("ceo/ceoProfile.jsp").forward(request, response);
             return;
         }
-
         if (image != null) {
-            String imgPath = aDao.getStaffById(id).getImage();
-            if (!imgPath.equals("assets/images/default.jpg")) {
+            String imgPath = currentAccount.getImage();
+            if (!imgPath.equals("assets/images/default.jpg") && !image.equals(imgPath)) {
                 deleteFile(imgPath);
             }
         } else {
-            image = aDao.getStaffById(id).getImage();
+            image = currentAccount.getImage();
         }
-
-        Staff staff = new Staff(id, username, "", image, email, firstName, lastName, gender, date, phone, address, bigDecimalValue,
-                id, LocalDateTime.MIN, new Role(aDao.getRoleIdByName(role), role));// Cập nhật thông tin người dùng
-        aDao.updateStaffInfo(staff, id);
-        response.sendRedirect("staffManagement");
+        currentAccount.setUsername(username);
+        currentAccount.setFirstname(firstName);
+        currentAccount.setLastname(lastName);
+        currentAccount.setAddress(address);
+        currentAccount.setGender(gender);
+        currentAccount.setImage(image);
+        aDao.updateStaffInfo(currentAccount, currentAccount.getId());
+        session.setAttribute("staff", currentAccount);
+        request.getRequestDispatcher("ceo/ceoProfile.jsp").forward(request, response);
     }
 
     /**
